refactor(docs): drop resolveAppRoot and flatten content-source layout
