--- apps/docs/src/lib/content-source/index.ts
@@ -1,2 +1,0 @@
-export { ContentSource } from './content-source';
-export type { ContentEntry, ContentFrontmatter, OrderBy, ContentSourceConfig } from './content-source';
