+++ apps/docs/src/lib/docs-source.ts
@@ -1,8 +1,5 @@
 import { join } from 'node:path';
-import { ContentSource } from './content-source/content-source';
-import { resolveAppRoot } from './resolve-app-root';
-
-const appRoot = resolveAppRoot(import.meta.url);
+import { ContentSource } from './content-source';
 
 /** Public URL prefix for docs pages. */
 export const DOCS_ROOT = '/docs';
@@ -19,6 +16,10 @@ export const DOCS_GROUP_ORDER = [
 	'Examples',
 ] as const;
 
+export const DOCS_GROUP_ORDER_INDEX = new Map<string, number>(
+	DOCS_GROUP_ORDER.map((name, index) => [name, index]),
+);
+
 /** Preferred section order for `llms.txt` (first slug segment). */
 export const LLM_SECTION_ORDER = [
 	'getting-started',
@@ -30,11 +31,7 @@ export const LLM_SECTION_ORDER = [
 	'tools',
 ] as const;
 
-export const docsContentRoot = join(appRoot, 'src/content/docs');
-
 export const docsSource = new ContentSource({
-	contentRoot: docsContentRoot,
+	contentRoot: join(import.meta.dir, '..', 'content', 'docs'),
 	orderBy: 'order',
 });
-
-export type { ContentEntry, ContentFrontmatter } from './content-source/content-source';
